+++ src/main/java/roomescape/controller/RoomescapeController.java
@@ -3,13 +3,12 @@ package roomescape.controller;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Person;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
@@ -17,25 +16,14 @@ import roomescape.domain.Reservations;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.dto.ReservationResponseDto;
 
-@Controller
+@RestController
 public class RoomescapeController {
 
     private final Reservations reservations = new Reservations();
     private final AtomicLong reservationIndex = new AtomicLong(1);
     private final AtomicLong personIndex = new AtomicLong(1);
 
-    @GetMapping("/admin")
-    public String admin() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String reservation() {
-        return "admin/reservation-legacy";
-    }
-
     @GetMapping("/reservations")
-    @ResponseBody
     public List<ReservationResponseDto> readReservations() {
         return reservations.getReservations().stream()
             .map(ReservationResponseDto::from)
@@ -43,7 +31,6 @@ public class RoomescapeController {
     }
 
     @PostMapping("/reservations")
-    @ResponseBody
     public Reservation createReservations(
         @RequestBody ReservationRequestDto reservationRequestDto) {
         Person person = new Person(personIndex.getAndIncrement(), reservationRequestDto.name());
@@ -56,7 +43,6 @@ public class RoomescapeController {
     }
 
     @DeleteMapping("/reservations/{id}")
-    @ResponseBody
     public void deleteReservation(@PathVariable(name = "id") Long id) {
         reservations.deleteById(id);
     }
